lesson_5/pyproject.toml
[project]
name = "custom-agent-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.84.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "mcp>=1.3.0",
    "litellm>=1.74.0",
    "instructor>=1.7.0",
    "pydantic>=2.0.0",
    "aiomysql>=0.2.0",
    "httpx>=0.27.0",
]
